+++ src/main.tsx
@@ -3,11 +3,12 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import App from './components/App/App';
 import { registerServiceWorker } from './lib/registerServiceWorker';
 import './styles/global.scss';
 import './config/langConfig';
 
+// Create root
 ReactDOM.createRoot(document.getElementById('root') as ReactDOM.Container).render(
   <React.StrictMode>
     <App />