+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Added panel resizing
 
+### Changed
+
+- Refactor App component
+
 ## [1.0.0] - 2026-07-13
 
 ### Added